09-18 ScrapActivity.kt ScrapAdapter 관련 에러 수정
수정사항
1.
scrappedNewsAdapter = ScrappedNewsAdapter(viewModel = scrappedNewsViewModel)

- Adpater로 viewModel을 받도록 수정

2.
ScrappedNewsAdapter의 생성자를
기존 viewModel: NewsDeatilViewModel에서
ScrappedNewsViewModel 타입으로 변경

3.
ScrappedNewsViewModel의 내부
_isScrapped, toggleScrap코드 추가하여
스크랩 기능 구현
